requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
cryptography==2.3.1
enum34==1.1.6
future==0.16.0
idna==2.7
ipaddress==1.0.22
pycparser==2.19
python-telegram-bot==11.1.0
six==1.11.0
